Compiles and loads reachability library as a plugin
