+++ reachability_description/launch/panda_husky/panda_husky_load_reachability.launch.py
@@ -110,7 +110,8 @@ def generate_launch_description():
             robot_description,
             robot_description_semantic,
             {"chain_group_name": "arm_base"}, # arm
-            {"robot_name": "panda_husky"} 
+            {"robot_name": "panda_husky"},
+            {"plugin_name": "reachability_description::ReachGraphReuleaux"}
         ]
     )
