+++ reachability_plugins/reach_plugin_reuleaux/apps/test_cube_discretization.cpp
@@ -1,5 +1,5 @@
 
-#include <reachability_description/quaternion_discretization.h>
+#include <reach_plugin_reuleaux/quaternion_discretization.h>
 #include <stdio.h>
 #include<fstream>
