+++ reachability_plugins/reach_plugin_reuleaux/include/reach_plugin_reuleaux/quaternion_discretization.h
